split up package types so that mapchete.types has no internal dependencies
