+++ mapchete/config/base.py
@@ -13,6 +13,7 @@ from shapely.geometry import box
 from shapely.geometry.base import BaseGeometry
 from shapely.ops import unary_union
 
+from mapchete.bounds import Bounds
 from mapchete.config.models import ProcessConfig
 from mapchete.config.parse import (
     get_zoom_levels,
@@ -34,13 +35,14 @@ from mapchete.io import MPath, absolute_path
 from mapchete.geometry import reproject_geometry
 from mapchete.tile import BufferedTile, BufferedTilePyramid, snap_geometry_to_tiles
 from mapchete.timer import Timer
-from mapchete.types import Bounds, BoundsLike, MPathLike, ZoomLevels
+from mapchete.types import BoundsLike, MPathLike
 from mapchete.validate import (
     validate_bounds,
     validate_bufferedtilepyramid,
     validate_values,
     validate_zooms,
 )
+from mapchete.zoom_levels import ZoomLevels
 
 logger = logging.getLogger(__name__)
