+++ mapchete/io/vector/types.py
@@ -1,4 +1,16 @@
-from typing import TypedDict
+from typing import Iterator, List, Optional, Protocol, TypedDict
 
-# VectorFileSchema = TypedDict('VectorFileSchema', {'geometry': Union[str, Tuple[str, ...]], 'properties': dict})
+from mapchete.types import GeoJSONLikeFeature
+from mapchete.types import BoundsLike, CRSLike
+
 VectorFileSchema = TypedDict("VectorFileSchema", {"geometry": str, "properties": dict})
+
+
+class FeatureCollectionProtocol(Protocol):
+    crs: CRSLike
+
+    def filter(
+        self, bounds: Optional[BoundsLike] = None, bbox: Optional[BoundsLike] = None
+    ) -> List[GeoJSONLikeFeature]: ...
+
+    def __iter__(self) -> Iterator[GeoJSONLikeFeature]: ...
