+++ mapchete/stac.py
@@ -12,7 +12,7 @@ from mapchete.io import MPath
 from mapchete.io.raster import write_raster_window
 from mapchete.geometry import reproject_geometry
 from mapchete.tile import BufferedTilePyramid
-from mapchete.types import Bounds
+from mapchete.bounds import Bounds
 
 logger = logging.getLogger(__name__)
